Small tweaks
- Fixed block destruction / map synchronization. Explosions are still broken though
- Tanks, bullets and everything now face east by default to avoid weird pi/2 summands or sin/cos swaps
- Consequentially, bullets are now launched with proper angle
- Fixed jitter caused by network sync and fixed physics timestep
- CTF flag now added as child to player tank when being carried
- Physics component now listens for position-update event to move itself instead of comparing transformations
- Physics component don't affect transform when body is static
- Loading error buttons now have cursor:pointer
- Generalized client-side game mode controller code. Yet to do the same with the huds
- Fixed CTF events visualization
- Added freeroam (FR) mode for (mostly) testing purposes

+++ src/entity/components/network/map/map-receiver.ts
@@ -10,7 +10,7 @@ export default class MapReceiver extends ReceiverComponent {
         receiveComponent.commandHandlers.set(Commands.BLOCK_UPDATE_COMMAND, (buffer) => {
             let x = buffer.readUint16()
             let y = buffer.readUint16()
-            let id = buffer.readUint8()
+            let id = buffer.readUint16()
             let damage = buffer.readFloat32()
 
             let Block = BlockState.getBlockStateClass(id)
